requirements.txt
Django==1.6.5
dj-database-url==0.2.2
dj-static==0.0.5
gunicorn==18.0
djangorestframework==2.3.10
static==0.4
wsgiref==0.1.2
psycopg2==2.5.1
django-registration==1.0
django-mail-queue==2.1.2
django-extensions==1.3.3
ipython==1.2.1
South==0.8.4
django-countries==2.1.2
django-rosetta==0.7.4
python-social-auth==0.1.23
Pillow==2.4.0
sorl-thumbnail==11.12
django-bootstrap-form==3.1
django-haystack==2.1.0
django-activity-stream==0.4.5beta1
django-mptt==0.6.0
elasticsearch==1.0.0
pyelasticsearch==0.6.1
django-contrib-comments==1.5
django-threadedcomments==0.9.0
django-sekizai==0.7